extract animation into function
